# packages/client/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rhizome rNet API",
    "version": "0.1.0",
    "description": "The rNet-compatible API exposed by the Rhizome store."
  },
  "servers": [{ "url": "/", "description": "Current Rhizome store" }],
  "paths": {
...
    "/rnet/v0/vibes/{id}/objects": {
      "get": {
        "operationId": "listVibeMediaObjects",
        "tags": ["Media Objects"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["mediaObjects"],
                  "properties": {
                    "mediaObjects": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MediaObject" }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      },
      "post": {
        "operationId": "addVibeMediaObjects",
        "tags": ["Media Objects"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["objects"],
                "properties": {
                  "objects": {
                    "type": "array",
                    "minItems": 1,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "pattern": "^rnet://object/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": { "description": "No content" },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      },
      "delete": {
        "operationId": "removeVibeMediaObjects",
        "tags": ["Media Objects"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["objects"],
                "properties": {
                  "objects": {
                    "type": "array",
                    "minItems": 1,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "pattern": "^rnet://object/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": { "description": "No content" },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
    "/rnet/v0/objects": {
      "post": {
        "operationId": "createMediaObjects",
        "tags": ["Media Objects"],
        "security": [{ "BearerAuth": [] }],
        "x-rhizome-auth": "user_or_client",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["metadata"],
                "properties": {
                  "metadata": {
                    "type": "string",
                    "contentMediaType": "application/json",
                    "contentSchema": {
                      "type": "object",
                      "required": ["objects"],
                      "properties": {
                        "vibe": {
                          "type": "string",
                          "pattern": "^rnet://vibe/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                        },
                        "objects": {
                          "type": "array",
                          "minItems": 1,
                          "items": {
                            "type": "object",
                            "properties": {
                              "elements": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string",
                                      "pattern": "^rnet://element/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                                    },
                                    {
                                      "type": "object",
                                      "required": ["upload", "kind", "mime"],
                                      "properties": {
                                        "upload": { "type": "string", "minLength": 1 },
                                        "kind": {
                                          "description": "Consumption strategy. Closed set — new kinds only by protocol revision.",
                                          "enum": ["text", "image", "audio", "video", "document"]
                                        },
                                        "mime": {
                                          "type": "string",
                                          "description": "IANA media type. mime is truth; kind is the consumption hint.",
                                          "pattern": "^[a-z]+/[a-zA-Z0-9][a-zA-Z0-9!#$&^_.+-]*$"
                                        }
                                      },
                                      "additionalProperties": false
                                    }
                                  ]
                                }
                              }
                            }
                          }
                        }
                      },
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["mediaObjects"],
                  "properties": {
                    "mediaObjects": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MediaObject" }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "403": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "415": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/objects/{id}": {
      "get": {
        "operationId": "getMediaObject",
        "tags": ["Media Objects"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MediaObject" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/objects/{id}/user": {
      "patch": {
        "operationId": "setMediaObjectUser",
        "tags": ["Media Objects"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["properties"],
                "properties": { "properties": { "type": "object" } },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MediaObject" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/rnet/v0/objects/{id}/inferred": {
      "put": {
        "operationId": "setMediaObjectInferred",
        "tags": ["Media Objects"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task", "entry"],
                "properties": {
                  "task": { "type": "string", "minLength": 1 },
                  "entry": {
                    "type": "object",
                    "required": ["model", "properties"],
                    "properties": {
                      "model": { "type": "string", "minLength": 1 },
                      "inferred_at": { "type": "string", "format": "date-time" },
                      "durable": {
                        "type": "boolean",
                        "default": false,
                        "description": "When true, a push task MUST NOT replace this entry. Durable entries hold understanding that accumulated rather than being computed from source — a user correction, a pattern an agent noticed across several objects. A task can never set this on its own output: durable means 'cannot be reproduced by re-running', and task output is by definition what re-running produces. Only agent runs and user-driven writes."
                      },
                      "properties": { "type": "object" },
                      "confidence": { "type": "number", "minimum": 0, "maximum": 1 }
                    },
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MediaObject" } }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "BearerAuth": { "type": "http", "scheme": "bearer" } },
    "schemas": {
...
      "IngestRecord": {
        "title": "IngestRecord",
        "description": "How an object came to be — the determinism disclosure. Consumers MAY gate on this record; the protocol guarantees the disclosure, not the trustworthiness.",
        "type": "object",
        "required": ["method", "reproducible"],
        "properties": {
          "method": {
            "description": "How this object was produced, named for what did the work. parser: a committed, versioned parser ran. generated_parser: an agent wrote a one-off parser, which ran — parser_hash pins it. agent: an agent extracted directly, writing no parser. authored: a person created the object in a client; nothing was parsed. Determinism runs parser > generated_parser > agent — generated code is hash-pinned and re-runnable, while freehand extraction is not. authored sits outside the ladder: nothing was derived.",
            "enum": ["parser", "generated_parser", "agent", "authored"]
          },
          "skill": {
            "description": "Skill identity + semver, when a corpus skill guided the run.",
            "type": ["string", "null"],
            "pattern": "^[a-z0-9][a-z0-9._-]*@\\d+\\.\\d+\\.\\d+([-+][0-9A-Za-z.-]+)?$"
          },
          "model": {
            "description": "Model identity, when a model participated in parsing.",
            "type": ["string", "null"],
            "minLength": 1
          },
          "parser_hash": {
            "description": "Content hash of generated parser code. Regeneration is a new parser, new provenance.",
            "type": ["string", "null"],
            "pattern": "^sha256:[a-f0-9]{64}$"
          },
          "reproducible": {
            "description": "Whether re-running the same method against source.origins yields identical output (modulo timestamps).",
            "type": "boolean"
          }
        },
        "allOf": [
          {
            "description": "parser_hash is REQUIRED when an agent generated the parser.",
            "if": {
              "required": ["method"],
              "properties": { "method": { "const": "generated_parser" } }
            },
            "then": {
              "required": ["parser_hash"],
              "properties": { "parser_hash": { "type": "string" } }
            }
          },
          {
            "description": "Freehand agent extraction is never reproducible.",
            "if": { "required": ["method"], "properties": { "method": { "const": "agent" } } },
            "then": { "properties": { "reproducible": { "const": false } } }
          }
        ],
        "additionalProperties": false
      },
...
      "MediaObject": {
        "title": "MediaObject",
        "description": "The core unit of meaning: properties plus zero or more MediaElements, carrying three property blocks with different mutation rights. Fields are values, elements are files — if you would query on it, it is a field; if you would open it, it is an element. Pure meaning-objects legitimately carry zero elements.",
        "type": "object",
        "required": ["rnet_schema", "uri", "owner", "type", "elements", "source"],
        "properties": {
          "rnet_schema": {
            "const": "0.1",
            "description": "Protocol version this document conforms to."
          },
          "uri": {
            "type": "string",
            "description": "UUIDv7-identified record.",
            "pattern": "^rnet://object/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "owner": {
            "type": "string",
            "description": "Immutable owner identity assigned by the store at creation. Ownership governs administration, not delegated access.",
            "pattern": "^rnet://id/[A-Za-z0-9._~-]+$"
          },
          "type": {
            "type": "string",
            "description": "Open vocabulary. The registered core vocabulary currently includes transaction and track; unregistered types such as post, photo, note, contact, event, book, article, and receipt remain legal.",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[a-z][a-z0-9_.-]*$"
          },
          "elements": {
            "type": "array",
            "description": "Ordered MediaElement URIs. MAY be empty.",
            "items": {
              "type": "string",
              "pattern": "^rnet://element/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          "keys": {
            "type": "object",
            "description": "External global identifiers for cross-service joins: isrc, isbn, fitid, url, ean, etc.",
            "additionalProperties": { "type": "string", "maxLength": 512 }
          },
          "source": {
            "type": "object",
            "description": "Written by ingestion runtimes only. Immutable after ingest — re-ingestion creates a new revision, never edits in place.",
            "required": ["ingest", "origins", "properties"],
            "properties": {
              "ingest": { "$ref": "#/components/schemas/IngestRecord" },
              "origins": {
                "type": "array",
                "minItems": 1,
                "description": "What this object was derived from — at least one, always. Either an origin artifact (rnet://origin/{uuid}) for ingested objects, or a client (rnet://client/{uuid}) for objects authored directly in an application. A filename string is not provenance.",
                "items": {
                  "type": "string",
                  "pattern": "^rnet://(origin|client)/[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "retrieved_at": { "type": "string", "format": "date-time" },
              "properties": {
                "type": "object",
                "description": "Facts the source format defines. Validated against the registered type vocabulary when one exists."
              }
            },
            "additionalProperties": false
          },
          "user": {
            "type": "object",
            "description": "Written by the owner via clients holding write:user. Freely mutable; revision-protected.",
            "required": ["properties"],
            "properties": {
              "properties": { "type": "object" },
              "updated_at": { "type": "string", "format": "date-time" }
            },
            "additionalProperties": false
          },
          "inferred": {
            "type": "object",
            "description": "A map keyed by writer and task: every key is {writer}:{task}. Memory scoped to this record — some entries are task output recomputed from source, others accumulated from corrections and agent observation and cannot be re-derived. A re-run replaces only its own key, and never a durable entry. Consumers MUST treat entries as advisory.",
            "propertyNames": {
              "pattern": "^(?:[a-z][a-z0-9._-]*|user/[A-Za-z0-9._~-]+):[a-z][a-z0-9_]*$"
            },
            "additionalProperties": {
              "type": "object",
              "required": ["model", "properties"],
              "properties": {
                "model": { "type": "string", "minLength": 1 },
                "inferred_at": { "type": "string", "format": "date-time" },
                "durable": {
                  "type": "boolean",
                  "default": false,
                  "description": "When true, a push task MUST NOT replace this entry. Durable entries hold understanding that accumulated rather than being computed from source — a user correction, a pattern an agent noticed across several objects. A task can never set this on its own output: durable means 'cannot be reproduced by re-running', and task output is by definition what re-running produces. Only agent runs and user-driven writes."
                },
                "properties": { "type": "object" },
                "confidence": { "type": "number", "minimum": 0, "maximum": 1 }
              },
              "additionalProperties": false
            }
          }
        },
        "patternProperties": {
          "^x-[a-z0-9-]+$": {
            "description": "Namespaced extensions. Consumers MUST ignore namespaces they don't understand."
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string", "minLength": 1 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string" },
          "code": { "type": "string", "minLength": 1 }
        },
        "additionalProperties": true
      }
    }
  }
}